requirements.txt
fastapi
uvicorn
pymongo[srv]==3.12
python-jose[cryptography]
passlib[bcrypt]
python-multipart